tpm2: prepare to add a proper tpm2 platform
This prepares things to add a proper tpm2 KeyData platform by doing the
following:
- Moves the code related to updating PCR policies to a separate file.
- Moves soon-to-be-deprecated code and APIs to new files.
- Refactors SealedKeyObject.UnsealFromTPM to be a wrapper around a new
internal function (SealedKeyObject.unsealDataFromTPM). The external API
will be deprecated and the new internal function will be shared with the
new code.
